+++ src/lib.rs
@@ -1,6 +1,7 @@
 //! Simple Verlet points and sticks implementation for bevy.
 //!
-//! If you are looking for cloth physics, please check [`bevy_silk`](https://github.com/ManevilleF/bevy_silk) instead,
+//! If you are looking for cloth physics, please check
+//! [`bevy_silk`](https://github.com/ManevilleF/bevy_silk) instead
 //!
 //! ## Bevy versions
 //!
@@ -12,6 +13,7 @@
 //!  | 0.4.x         | 0.8.x     |
 //!  | 0.5.x         | 0.9.x     |
 //!  | 0.6.x         | 0.11.x    |
+//!  | 0.7.x         | 0.12.x    |
 //!
 //! ## Features
 //!